requirements.txt
httpx==0.27.0
rich==13.7.1
python-dotenv==1.0.1
streamlit==1.35.0
reportlab==4.4.10
